add adopt

+++ src/main/java/com/project/soshuceapi/repositories/ActionLogDetailRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActionLogDetailRepository extends JpaRepository<ActionLogDetail, Integer> {
+public interface ActionLogDetailRepository extends JpaRepository<ActionLogDetail, Long> {
 }
